Macro crepe::crepe

crepe!() { /* proc-macro */ }

The main macro, which lets you write a Datalog program declaratively.

Example

A program to calculate transitive closure might be written as:

mod datalog {
    use crepe::crepe;

    crepe! {
        @input
        struct Edge(i32, i32);

        @output
        struct Tc(i32, i32);

        Tc(x, y) <- Edge(x, y);
        Tc(x, z) <- Edge(x, y), Tc(y, z);
    }

    pub fn run(edges: &[(i32, i32)]) -> Vec<(i32, i32)> {
        let mut runtime = Crepe::new();
        runtime.extend(edges.iter().map(|&(a, b)| Edge(a, b)));
        let (tc,) = runtime.run();
        tc.into_iter().map(|Tc(a, b)| (a, b)).collect()
    }
}

Generated Code

Each struct in the program is turned into a Datalog relation, while each line of the form goal <- clause1, clause2, ...; or fact; defines a logic programming rule that is used to derive new relations.

In addition to the relation structs, the macro also defines a Crepe struct representing the runtime. This is the primary way that you interact with Crepe. It provides a couple methods and traits (here Rel is a placeholder for the name of your relation):

Crepe::new(): construct a new runtime

Implements std::iter::Extend<Rel> for each @input struct.

Crepe::extend(&mut self, iter: impl IntoIterator<Item = Rel>)

Implements std::iter::Extend<&'a Rel> for each @input struct.

Crepe::extend(&mut self, iter: impl IntoIterator<Item = &'a Rel>)

Crepe::run(self): evaluates the Datalog program on the given inputs, consuming the runtime object, and returns a tuple of HashSet<Rel>s containing the final derived @output structs.

Crepe::run_with_hasher::<S: BuildHasher + Default>(self): similar to the run method, but internally uses a custom hasher.

In order for the engine to work, all relations must be tuple structs, and they automatically derive the Eq, PartialEq, Hash, Copy, and Clone traits. This is necessary in order to create efficient indices that are used in Datalog evaluation. If you want to use Crepe with types that do not implement Copy, consider passing references instead.

Datalog Syntax Extensions

This library supports arbitrary Rust expression syntax within rules for constructing new relations, as well as Boolean expressions that are evaluated directly as clauses in rules. Basically, this gives you seamless Rust interoperability within your Datalog programs.

For instance, here is a program that calculates the first 25 Fibonacci numbers using arithmetic functors:

crepe! {
    @output
    struct Fib(u32, u32);

    Fib(0, 0) <- (true);
    Fib(1, 1); // shorthand for `Fib(1, 1) <- (true);`

    Fib(n + 2, x + y) <- Fib(n, x), Fib(n + 1, y), (n + 2 <= 25);
}

Note that all Boolean conditions within the clauses of rules are evaluated in-place, and they must be surrounded by parentheses.

We also support let-bindings in rules, including bindings that destructure their arguments conditionally. See tests/test_destructure.rs for an example of this.

crepe! {
    @input
    struct Value<'a>(&'a str);

    @output
    struct Squared(i32, i32);

    Squared(n, x) <-
        Value(string),
        let Ok(n) = string.parse(),
        let x = n * n;
}

The last built-in control flow feature is iteration over data. Rules can enumerate values from an iterator, allowing them to use data from outside of Crepe without having to convert functions to use work-arounds. For example, to access the characters of a string, you could write:

crepe! {
    @input
    struct Name<'a>(&'a str);

    @output
    struct NameContainsLetter<'a>(&'a str, char);

    NameContainsLetter(name, letter) <- Name(name), for letter in name.chars();
}

Evaluation Mode

All generated code uses semi-naive evaluation (see Chapter 3 of Datalog and Recursive Query Processing), and it is split into multiple strata to enable stratified negation. For example, we can extend the first example to also compute the complement of transitive closure in a graph:

mod datalog {
    use crepe::crepe;

    crepe! {
        @input
        struct Edge(i32, i32);

        @output
        struct Tc(i32, i32);

        struct Node(i32);

        @output
        struct NotTc(i32, i32);

        Tc(x, y) <- Edge(x, y);
        Tc(x, z) <- Edge(x, y), Tc(y, z);

        Node(x) <- Edge(x, _);
        Node(x) <- Edge(_, x);
        NotTc(x, y) <- Node(x), Node(y), !Tc(x, y);
    }

    pub fn run(edges: &[(i32, i32)]) -> (Vec<(i32, i32)>, Vec<(i32, i32)>) {
        let mut runtime = Crepe::new();
        runtime.extend(edges.iter().map(|&(a, b)| Edge(a, b)));
        let (tc, not_tc) = runtime.run();
        (
            tc.into_iter().map(|Tc(a, b)| (a, b)).collect(),
            not_tc.into_iter().map(|NotTc(a, b)| (a, b)).collect(),
        )
    }
}

Lifetimes and Attributes

This macro allows you to specify attributes, visibility modifiers, and lifetimes on the relation structs. These can include additional derive attributes, and lifetimes can be used to construct relations that include non-owning references. The following example computes suffixes of words.

use crepe::crepe;

crepe! {
    @input
    struct Word<'a>(&'a str);

    @output
    #[derive(Debug)]
    struct Suffix<'a>(&'a str);

    Suffix(w) <- Word(w);
    Suffix(&w[1..]) <- Suffix(w), (!w.is_empty());
}

fn main() {
    let mut runtime = Crepe::new();
    runtime.extend(&[Word("banana"), Word("bandana")]);
    let (suffixes,) = runtime.run();
    println!("{:?}", suffixes);
}

We also support the ref keyword for binding a variable by reference rather than copying it, which can improve performance in some cases.

Hygiene

In addition to the relation structs, this macro generates implementations of a private struct named Crepe for the runtime. Therefore, it is recommended to place each Datalog program within its own module, to prevent name collisions.